+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.7.5] - 2020-07-02
+### Fixed
+- Iron/gold smelting/blasting recipes missing [#90]
+
+## [0.7.4] - 2020-06-30
 ### Changed
 - Switched over to Forge's config system. Configs may get reset.
 ### Fixed